:sparkles: Implement recording the 'current' version
Version recording is a heuristic where we grab the version from the
settings/project (this will be made dynamic later), and combine it with
the machine name. Only if no similar combination was recorded in the
past hour do we record it (again).

This 'debounce' mechanism is not meant to be exact, rather just
account for a containerized landscape where machines can be spun up
frequently or pods may be re-scheduled and/or restarted due to
health checks failing.

We *assume* that machines have their clocks properly sychronized via
NTP to account for (large) clock drift, as we order the deployed
versions on a timestamp that's generated by each machine/container.
Some drift (seconds/minutes) should not be a problem either, if
there are multiple versions recorded close together in time it's
likely that they'll be the same version due to rolling release
upgrades. Otherwise, any crashes or odd things showing up will be
an indication that something is probably wrong.