:sparkles: Define upgrade checks in terms of allowed version ranges
Currently the UpgradeCheck class supports specifying 'current' version ranges
that enable upgrading to a particular version (todo). The check verifies that
the current application version is contained in any of the provided ranges.

You can pass a single version range instance for simplicity, or a collection
of version ranges for more complicated scenario's.

In the future, additional checks such as running management commands and/or
arbitrary check scripts/callbacks will be possible to be specified in the
upgrade check too.