:zap: Calculate parsed version once
The parsed SemVer version is now calculated at init time and only once,
which results in faster developer feedback and less object
instantiations if the range is checked multiple times.

For minimal memory usage, slots are leveraged and unsafe hash enabled
because the objects are not supposed to be mutable.